EbmlElement: use size_t for HeadSize()
EbmlID::Length is a size_t so the operation ends up being a size_t that is downsized
to a uint8. It probably has negative performance.

Fixes warning in MSVC.

+++ ebml/EbmlElement.h
@@ -433,7 +433,7 @@ class EBML_DLL_API EbmlElement {
     virtual bool IsDummy() const {return false;}
     virtual bool IsMaster() const {return false;}
 
-    uint8 HeadSize() const {
+    size_t HeadSize() const {
       return EBML_ID_LENGTH((const EbmlId&)*this) + CodedSizeLength(Size, SizeLength, bSizeIsFinite);
     } /// return the size of the head, on reading/writing
